# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentação da API",
        "description": "API desenvolvida com Laravel Sanctum e documentada via Swagger",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Ambiente local"
        }
    ],
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Realiza login e retorna o token de acesso",
                "operationId": "c373bc4cffa4a56fc64e3979b8ee7b96",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "admin@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "senha"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login realizado com sucesso, retorna o token"
                    },
                    "401": {
                        "description": "Credenciais inválidas"
                    }
                },
                "security": []
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Faz logout e revoga o token",
                "operationId": "5ff2dc8b48bff52908d3821a0ecc9b3e",
                "responses": {
                    "200": {
                        "description": "Logout realizado com sucesso"
                    },
                    "401": {
                        "description": "Não autorizado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/fotos": {
            "post": {
                "tags": [
                    "FotoPessoa"
                ],
                "summary": "Fazer upload de uma ou mais fotos para uma pessoa",
                "description": "Upload de múltiplas fotos associadas a uma pessoa.",
                "operationId": "uploadFotosPessoa",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "pes_id",
                                    "fotos"
                                ],
                                "properties": {
                                    "pes_id": {
                                        "description": "ID da pessoa",
                                        "type": "integer"
                                    },
                                    "fotos": {
                                        "description": "Arquivos de imagem para upload",
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Fotos salvas com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/FotoPessoa"
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação"
                    },
                    "401": {
                        "description": "Não autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/pessoas": {
            "get": {
                "tags": [
                    "Pessoa"
                ],
                "summary": "Listar pessoas",
                "operationId": "7c7d2c9decc48109bcd470035e629ecb",
                "parameters": [
                    {
                        "name": "expira_em",
                        "in": "query",
                        "description": "Tempo de expiração da URL da imagem em minutos",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de pessoas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Pessoa"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Pessoa"
                ],
                "summary": "Cadastrar uma nova pessoa",
                "operationId": "b72258a0649966eb7cb39a9fdc67bf4b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Pessoa"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Pessoa criada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Pessoa"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro de validação"
                    },
                    "401": {
                        "description": "Não autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/pessoas/{id}": {
            "get": {
                "tags": [
                    "Pessoa"
                ],
                "summary": "Exibir uma pessoa pelo ID",
                "operationId": "9767860b5875b66fc3535a00d47aec2d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da pessoa",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "expira_em",
                        "in": "query",
                        "description": "Tempo de expiração da URL da imagem em minutos",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pessoa encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Pessoa"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Ação não permitida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "This action is unauthorized."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Pessoa não encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No query results for model [Pessoa] 1"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Pessoa"
                ],
                "summary": "Atualizar uma pessoa pelo ID",
                "operationId": "da7d1255c980a411d322994cbca98632",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da pessoa",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Pessoa"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Pessoa atualizada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Pessoa"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Não autorizado"
                    },
                    "404": {
                        "description": "Pessoa não encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Pessoa"
                ],
                "summary": "Excluir uma pessoa pelo ID",
                "operationId": "6ac497549c4c86a2d7bbd144d43dffb3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da pessoa",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Pessoa excluída com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Ação não permitida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "This action is unauthorized."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Pessoa não encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No query results for model [Pessoa] 1"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "FotoPessoa": {
                "title": "FotoPessoa",
                "required": [
                    "id",
                    "pes_id",
                    "url",
                    "url_temporaria"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "pes_id": {
                        "type": "integer",
                        "example": 12
                    },
                    "url": {
                        "type": "string",
                        "format": "uri",
                        "example": "https://minio.local/fotos_pessoas/12/foto1.jpg"
                    },
                    "url_temporaria": {
                        "type": "string",
                        "format": "uri",
                        "example": "http://localhost:9000/fotos_pessoas/12/foto1.jpg?X-Amz-Expires=600&X-Amz-Signature=abc123..."
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-04-10T14:48:00.000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-04-10T14:50:00.000Z"
                    }
                },
                "type": "object"
            },
            "Pessoa": {
                "title": "Pessoa",
                "properties": {
                    "pes_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "pes_nome": {
                        "type": "string",
                        "example": "Ana Maria"
                    },
                    "pes_data_nascimento": {
                        "type": "string",
                        "format": "date",
                        "example": "1986-02-08"
                    },
                    "idade": {
                        "type": "integer",
                        "example": 39
                    },
                    "pes_sexo": {
                        "type": "string",
                        "example": "Feminino"
                    },
                    "pes_mae": {
                        "type": "string",
                        "example": "Mãe da Ana Maria"
                    },
                    "pes_pai": {
                        "type": "string",
                        "example": "Pai da Ana Maria"
                    },
                    "fotos": {
                        "type": "array",
                        "items": {
                            "type": "object"
                        }
                    },
                    "servidores_efetivos": {
                        "type": "array",
                        "items": {
                            "type": "object"
                        }
                    },
                    "servidores_temporarios": {
                        "type": "array",
                        "items": {
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
            "ServidorEfetivo": {
                "title": "Servidor Efetivo",
                "properties": {
                    "pes_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "se_matricula": {
                        "type": "string",
                        "example": "20241234"
                    }
                },
                "type": "object"
            },
            "ServidorTemporario": {
                "title": "Servidor Temporário",
                "properties": {
                    "pes_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "st_data_admissao": {
                        "type": "string",
                        "format": "date",
                        "example": "2023-05-01"
                    },
                    "st_data_demissao": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-04-01"
                    }
                },
                "type": "object"
            },
            "User": {
                "title": "User",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Admin"
                    },
                    "email": {
                        "type": "string",
                        "example": "admin@email.com"
                    },
                    "role": {
                        "type": "string",
                        "example": "admin"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Autenticação",
            "description": "Autenticação"
        },
        {
            "name": "FotoPessoa",
            "description": "FotoPessoa"
        },
        {
            "name": "Pessoa",
            "description": "Pessoa"
        }
    ],
    "security": {
        "1": {
            "sanctum": []
        }
    }
}